Range type changed to double

+++ app/src/main/java/ru/orlovvv/ort/models/LocationPreview.kt
@@ -10,7 +10,7 @@ data class LocationPreview(
     val _id: String,
     val address: String,
     val name: String,
-    val range: Int,
+    val range: Double,
     val services: List<String>,
     val stars: Int
 ) : Serializable
